Rename packagelimits_t to powerlimits_t and use correct source for TDP.
Aparently the definition of the maximum power consumption (TDP) differs
between CPUs and even systems. In theory it should be enough ti query
the nominal power, the thermal_spec_power field. But on some systems
maximum_power is even higher and the CPU exceeds thermal_spec_power. On
other systems maximum_power is set to 0. Choose the highest value.